refactor: Refactor mergeHalves function for efficiency
Function originally used shift to get elements for comparison. Since
shift has to re-index all array elements, it becomes very inefficient
for large datasets.

To fix this issue, restructure the while loop into a for loop, that
loops for the total number of items in the left and right arrays. Then,
use two 'index' variables (leftIdx and rightIdx) to track which items
are being compared and move through the array.